+++ agb/src/sound/mixer/sw_mixer.rs
@@ -1,5 +1,6 @@
 use core::cell::RefCell;
 use core::marker::PhantomData;
+use core::num::NonZero;
 
 use alloc::boxed::Box;
 use alloc::vec::Vec;
@@ -114,7 +115,7 @@ pub struct Mixer<'gba> {
 
     buffer: raw_box::RawBoxDrop<MixerBuffer, InternalAllocator>,
     channels: [Option<SoundChannel>; 8],
-    indices: [i32; 8],
+    indices: [u32; 8],
     frequency: Frequency,
 
     working_buffer: Box<[Num<i16, 4>], InternalAllocator>,
@@ -150,7 +151,8 @@ pub struct Mixer<'gba> {
 /// mixer.channel(&bgm_channel_id).expect("Expected to still be playing").stop();
 /// # }
 /// ```
-pub struct ChannelId(usize, i32);
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub struct ChannelId(usize, NonZero<u32>);
 
 impl Mixer<'_> {
     pub(super) fn new(frequency: Frequency) -> Self {
@@ -301,8 +303,9 @@ impl Mixer<'_> {
             }
 
             channel.replace(new_channel);
-            self.indices[i] += 1;
-            return Some(ChannelId(i, self.indices[i]));
+            let generation = self.next_index(i);
+
+            return Some(ChannelId(i, generation));
         }
 
         if new_channel.priority == SoundPriority::Low {
@@ -315,11 +318,21 @@ impl Mixer<'_> {
             }
 
             channel.replace(new_channel);
-            self.indices[i] += 1;
-            return Some(ChannelId(i, self.indices[i]));
+            let generation = self.next_index(i);
+
+            return Some(ChannelId(i, generation));
         }
 
         panic!("Cannot play more than 8 sounds at once");
+    }
+
+    fn next_index(&mut self, i: usize) -> NonZero<u32> {
+        self.indices[i] = self.indices[i].wrapping_add(1);
+        if self.indices[i] == 0 {
+            self.indices[i] += 1;
+        }
+
+        NonZero::new(self.indices[i]).expect("Should be bigger than 0")
     }
 
     /// Lets you modify an already playing channel.
@@ -347,7 +360,7 @@ impl Mixer<'_> {
     /// ```
     pub fn channel(&mut self, id: &ChannelId) -> Option<&'_ mut SoundChannel> {
         if let Some(channel) = &mut self.channels[id.0]
-            && self.indices[id.0] == id.1
+            && self.indices[id.0] == id.1.into()
             && !channel.is_done
         {
             return Some(channel);
@@ -805,5 +818,10 @@ mod test {
             ]
         );
         assert_eq!(result, num!(7.0));
+    }
+
+    #[test_case]
+    fn channel_id_none_niche(_: &mut crate::Gba) {
+        assert_eq!(size_of::<Option<ChannelId>>(), size_of::<ChannelId>());
     }
 }